Change DB to Sqlite

+++ AMS.Model/Partials/AmsNeo4JNode.cs
@@ -10,7 +10,7 @@ using QOQNOS.Core;
 
 namespace AMS.Model.Models
 {
-    public partial class AmsNeo4JNode : IHaveId<int>
+    public partial class AmsNeo4JNode : IHaveId<long>
     {
         [JsonIgnore]
